Add a fantastically shitty workaround for The Great Corner Case of 2020.
(Namely, the 2019.12 -> 2020.01 KOReader update).

In short: at the end of the process, zsync decides that it's missing a
single block, but the computed needed ranges (2 of them, in our case)
repeatedly fail to fullfill that missing block, which makes zsync
loop forever on those two exact same ranges.

Just remember the amount of ranges requested in the last fetch,
and if it matches, flag this as a retry, and extend each range
by 4096 bytes per retry.

That happened to do the job in this case, no idea how sane this is.

Note that the ranges optimization happened to workaround the issue,
too, so, yay.